Fixed some logic to follow coding standards.
