--- logger/formatter_interface.py
@@ -1,11 +1,0 @@
-from abc import abstractmethod, ABCMeta
-
-
-class FormatterInterface(metaclass=ABCMeta):
-    @abstractmethod
-    def __init__(self) -> None:
-        pass
-
-    @abstractmethod
-    def format(self, message: str, log_level: str) -> str:
-        """format the message."""
