+++ logger/logger.py
@@ -1,9 +1,31 @@
+from abc import abstractmethod
 from typing import final
 
-from logger.formatter_interface import FormatterInterface
-from logger.logger_interface import LoggerInterface
+from logger.formatters import FormatterInterface
 
 LOG_LEVELS = ['critical', 'error', 'warning', 'info', 'debug']
+
+
+class LoggerInterface:
+    @abstractmethod
+    def critical(self, message: str) -> str:
+        """Logs critical message into specific output channel."""
+
+    @abstractmethod
+    def error(self, message: str) -> str:
+        """Logs error message into specific output channel."""
+
+    @abstractmethod
+    def warning(self, message: str) -> str:
+        """Logs warning message into specific output channel."""
+
+    @abstractmethod
+    def info(self, message: str) -> str:
+        """Logs info message into specific output channel."""
+
+    @abstractmethod
+    def debug(self, message: str) -> str:
+        """Logs debug message into specific output channel."""
 
 
 @final # pylint: disable=using-final-decorator-in-unsupported-version
